# docs/api/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Open Archiver API",
		"version": "1.0.0",
		"description": "REST API for Open Archiver — an open-source email archiving platform. All authenticated endpoints require a Bearer JWT token obtained from `POST /v1/auth/login`, or an API key passed as a Bearer token.",
		"license": {
			"name": "SEE LICENSE IN LICENSE",
			"url": "https://github.com/LogicLabs-OU/OpenArchiver/blob/main/LICENSE"
		},
		"contact": {
			"name": "Open Archiver",
			"url": "https://openarchiver.com"
		}
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Local development"
		}
	],
	"security": [
		{
			"bearerAuth": []
		},
		{
			"apiKeyAuth": []
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "JWT obtained from `POST /v1/auth/login`. Pass as `Authorization: Bearer <token>`."
			},
			"apiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "X-API-KEY",
				"description": "API key generated via `POST /v1/api-keys`. Pass as `X-API-KEY: <key>`."
			}
		},
		"responses": {
			"Unauthorized": {
				"description": "Authentication is required or the token is invalid/expired.",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorMessage"
						},
						"example": {
							"message": "Unauthorized"
						}
					}
				}
			},
...
			"InternalServerError": {
				"description": "An unexpected error occurred on the server.",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/ErrorMessage"
						},
						"example": {
							"message": "Internal server error"
						}
					}
				}
			}
		},
		"schemas": {
			"ErrorMessage": {
				"type": "object",
				"properties": {
					"message": {
						"type": "string",
						"description": "Human-readable error description.",
						"example": "An error occurred."
					}
				},
				"required": ["message"]
			},
...
			"SearchResults": {
				"type": "object",
				"properties": {
					"hits": {
						"type": "array",
						"description": "Array of matching archived email objects, potentially with highlighted fields.",
						"items": {
							"type": "object"
						}
					},
					"total": {
						"type": "integer",
						"example": 42
					},
					"page": {
						"type": "integer",
						"example": 1
					},
					"limit": {
						"type": "integer",
						"example": 10
					},
					"totalPages": {
						"type": "integer",
						"example": 5
					},
					"processingTimeMs": {
						"type": "integer",
						"description": "Meilisearch query processing time in milliseconds.",
						"example": 12
					}
				},
				"required": ["hits", "total", "page", "limit", "totalPages", "processingTimeMs"]
			},
...
		}
	},
	"paths": {
...
		"/v1/search": {
			"get": {
				"summary": "Search archived emails",
				"description": "Performs a full-text search across indexed archived emails using Meilisearch. Requires `search:archive` permission. Use `keywords` for global search, or one field-specific text parameter (`subject`, `body`, `attachmentFilename`, or `attachmentContent`). Requests with more than one text search parameter are rejected.",
				"operationId": "searchEmails",
				"tags": ["Search"],
				"security": [
					{
						"bearerAuth": []
					},
					{
						"apiKeyAuth": []
					}
				],
				"parameters": [
					{
						"name": "keywords",
						"in": "query",
						"required": false,
						"description": "Global search query string across searchable email fields.",
						"schema": {
							"type": "string",
							"example": "invoice Q4"
						}
					},
					{
						"name": "from",
						"in": "query",
						"required": false,
						"description": "Exact sender email address filter. Input is trimmed and lowercased.",
						"schema": {
							"type": "string",
							"format": "email",
							"example": "sender@example.com"
						}
					},
					{
						"name": "to",
						"in": "query",
						"required": false,
						"description": "Exact recipient email address filter. Input is trimmed and lowercased.",
						"schema": {
							"type": "string",
							"format": "email",
							"example": "recipient@example.com"
						}
					},
					{
						"name": "cc",
						"in": "query",
						"required": false,
						"description": "Exact CC recipient email address filter. Input is trimmed and lowercased.",
						"schema": {
							"type": "string",
							"format": "email",
							"example": "copy@example.com"
						}
					},
					{
						"name": "bcc",
						"in": "query",
						"required": false,
						"description": "Exact BCC recipient email address filter. Input is trimmed and lowercased.",
						"schema": {
							"type": "string",
							"format": "email",
							"example": "hidden@example.com"
						}
					},
					{
						"name": "subject",
						"in": "query",
						"required": false,
						"description": "Search only the email subject. Cannot be combined with `keywords` or other field-specific text parameters.",
						"schema": {
							"type": "string",
							"example": "quarterly report"
						}
					},
					{
						"name": "body",
						"in": "query",
						"required": false,
						"description": "Search only the email body. Cannot be combined with `keywords` or other field-specific text parameters.",
						"schema": {
							"type": "string",
							"example": "contract renewal"
						}
					},
					{
						"name": "attachmentFilename",
						"in": "query",
						"required": false,
						"description": "Search only attachment filenames. Cannot be combined with `keywords` or other field-specific text parameters.",
						"schema": {
							"type": "string",
							"example": "invoice.pdf"
						}
					},
					{
						"name": "attachmentContent",
						"in": "query",
						"required": false,
						"description": "Search only extracted attachment content. Cannot be combined with `keywords` or other field-specific text parameters.",
						"schema": {
							"type": "string",
							"example": "purchase order"
						}
					},
					{
						"name": "dateFrom",
						"in": "query",
						"required": false,
						"description": "Include emails sent on or after this date.",
						"schema": {
							"type": "string",
							"format": "date",
							"example": "2026-01-01"
						}
					},
					{
						"name": "dateTo",
						"in": "query",
						"required": false,
						"description": "Include emails sent on or before this date.",
						"schema": {
							"type": "string",
							"format": "date",
							"example": "2026-01-31"
						}
					},
					{
						"name": "page",
						"in": "query",
						"required": false,
						"description": "Page number for pagination.",
						"schema": {
							"type": "integer",
							"default": 1,
							"example": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"description": "Number of results per page.",
						"schema": {
							"type": "integer",
							"default": 10,
							"example": 10
						}
					},
					{
						"name": "matchingStrategy",
						"in": "query",
						"required": false,
						"description": "Meilisearch matching strategy. `last` returns results containing at least one keyword; `all` requires all keywords; `frequency` sorts by keyword frequency.",
						"schema": {
							"type": "string",
							"enum": ["last", "all", "frequency"],
							"default": "last"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Search results.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SearchResults"
								}
							}
						}
					},
					"400": {
						"description": "Search criteria are missing or invalid.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"500": {
						"$ref": "#/components/responses/InternalServerError"
					}
				}
			}
		},
...
	},
...
}